Refactored core, decorator and extension
