+++ tests/base_test.py
@@ -14,12 +14,8 @@ try:
 except ImportError:
     import unittest
 
-try:
-    # this is how you would normally import
-    from flask.ext.cors import *
-except:
-    # support local usage without installed package
-    from flask_cors import *
+from flask_cors import *
+from flask_cors.core import *
 
 
 class FlaskCorsTestCase(unittest.TestCase):
