+++ tests/test_exception_interception.py
@@ -10,12 +10,8 @@
 """
 from tests.base_test import FlaskCorsTestCase
 from flask import Flask, abort
-try:
-    # this is how you would normally import
-    from flask.ext.cors import *
-except:
-    # support local usage without installed package
-    from flask_cors import *
+from flask_cors import *
+from flask_cors.core import *
 
 
 def add_routes(app):
