+++ tests/test_vary_header.py
@@ -12,12 +12,8 @@
 from tests.base_test import FlaskCorsTestCase, AppConfigTest
 from flask import Flask, Response
 
-try:
-    # this is how you would normally import
-    from flask.ext.cors import *
-except:
-    # support local usage without installed package
-    from flask_cors import *
+from flask_cors import *
+from flask_cors.core import *
 
 
 class VaryHeaderTestCase(FlaskCorsTestCase):
